progress bar

+++ front_end/path_way_flu/lib/app/data/model/lession.dart
@@ -1,43 +1,3 @@
-// class Lession {
-//   String id;
-//   String creatorName;
-//   String subject;
-//   String profileImage;
-//   int countOfLesson;
-//   int watchTime;
-//   String title;
-//   String coverImage;
-//   List<String> lessonId;
-
-//   Lession({
-//     required this.id,
-//     required this.creatorName,
-//     required this.subject,
-//     required this.title,
-//     required this.coverImage,
-//     this.profileImage = '',
-//     this.countOfLesson = 0,
-//     this.watchTime = 0,
-//     this.lessonId = const [],
-//   });
-
-// factory Lession.fromJson(Map<String, dynamic> json) {
-//   return Lession(
-//     id: json['_id'] ?? '',
-//     creatorName: json['creatorName'] ?? '',
-//     subject: json['subject'] ?? '',
-//     coverImage: json['coverImage'] ?? '',
-//     profileImage: json['profileImage'] ?? '',
-//     countOfLesson: json['countOfLesson'] ?? 0,
-//     watchTime: json['watchTime'] ?? 0,
-//     title: json['title'] ?? '',
-//     lessonId: json['lessonId'] != null
-//         ? List<String>.from(json['lessonId'])
-//         : [],
-//   );
-// }
-// }
-
 class Lesson {
   String id;
   String creatorName;
@@ -45,7 +5,7 @@ class Lesson {
   String profileImage;
   String coverImage;
   int countOfLesson;
-  double price;
+  int price;
   String creatorId; // Changed to String
   int watchTime;
   String title;
